Extract exported enums from game commands
