+++ src/enums/option_types/answer_type.ts
@@ -1,0 +1,7 @@
+export enum AnswerType {
+    TYPING = "typing",
+    TYPING_TYPOS = "typingtypos",
+    MULTIPLE_CHOICE_EASY = "easy",
+    MULTIPLE_CHOICE_MED = "medium",
+    MULTIPLE_CHOICE_HARD = "hard",
+}
