+++ src/enums/option_types/artist_type.ts
@@ -1,0 +1,5 @@
+export enum ArtistType {
+    SOLOIST = "soloists",
+    GROUP = "groups",
+    BOTH = "both",
+}
