+++ src/enums/option_types/gender.ts
@@ -1,0 +1,6 @@
+export enum Gender {
+    MALE = "male",
+    FEMALE = "female",
+    COED = "coed",
+    ALTERNATING = "alternating",
+}
