+++ src/enums/option_types/guess_mode_type.ts
@@ -1,0 +1,5 @@
+export enum GuessModeType {
+    SONG_NAME = "song",
+    ARTIST = "artist",
+    BOTH = "both",
+}
