+++ src/enums/option_types/language_type.ts
@@ -1,0 +1,4 @@
+export enum LanguageType {
+    KOREAN = "korean",
+    ALL = "all",
+}
