+++ src/enums/option_types/multiguess_type.ts
@@ -1,0 +1,4 @@
+export enum MultiGuessType {
+    ON = "on",
+    OFF = "off",
+}
