+++ src/enums/option_types/ost_preference.ts
@@ -1,0 +1,5 @@
+export enum OstPreference {
+    INCLUDE = "include",
+    EXCLUDE = "exclude",
+    EXCLUSIVE = "exclusive",
+}
