+++ src/enums/option_types/release_type.ts
@@ -1,0 +1,4 @@
+export enum ReleaseType {
+    OFFICIAL = "official",
+    ALL = "all",
+}
