+++ src/enums/option_types/seek_type.ts
@@ -1,0 +1,5 @@
+export enum SeekType {
+    BEGINNING = "beginning",
+    RANDOM = "random",
+    MIDDLE = "middle",
+}
