+++ src/enums/option_types/shuffle_type.ts
@@ -1,0 +1,6 @@
+export enum ShuffleType {
+    RANDOM = "random",
+    WEIGHTED_EASY = "weighted_easy",
+    WEIGHTED_HARD = "weighted_hard",
+    POPULARITY = "popularity",
+}
