+++ src/enums/option_types/special_type.ts
@@ -1,0 +1,9 @@
+export enum SpecialType {
+    REVERSE = "reverse",
+    SLOW = "slow",
+    FAST = "fast",
+    FASTER = "faster",
+    LOW_PITCH = "lowpitch",
+    HIGH_PITCH = "highpitch",
+    NIGHTCORE = "nightcore",
+}
