+++ src/enums/option_types/subunit_preference.ts
@@ -1,0 +1,4 @@
+export enum SubunitsPreference {
+    INCLUDE = "include",
+    EXCLUDE = "exclude",
+}
