+++ src/test/ci/game_utils.test.ts
@@ -7,12 +7,12 @@ import {
     getAvailableSongCount,
     getMatchingGroupNames,
 } from "../../helpers/game_utils";
-import { SubunitsPreference } from "../../commands/game_options/subunits";
 import { state } from "../../kmq_worker";
 import GameSession from "../../structures/game_session";
-import { OstPreference } from "../../commands/game_options/ost";
-import { ReleaseType } from "../../commands/game_options/release";
 import * as discordUtils from "../../helpers/discord_utils";
+import { ReleaseType } from "../../enums/option_types/release_type";
+import { OstPreference } from "../../enums/option_types/ost_preference";
+import { SubunitsPreference } from "../../enums/option_types/subunit_preference";
 
 const sandbox = sinon.createSandbox();
