+++ src/test/ci/song_selector.test.ts
@@ -1,22 +1,20 @@
 import assert from "assert";
 import sinon from "sinon";
-import { Gender } from "../../commands/game_options/gender";
-import { OstPreference } from "../../commands/game_options/ost";
-import { ReleaseType } from "../../commands/game_options/release";
-import { SubunitsPreference } from "../../commands/game_options/subunits";
 import GuildPreference from "../../structures/guild_preference";
 import SongSelector, {
     LAST_PLAYED_SONG_QUEUE_SIZE,
 } from "../../structures/song_selector";
 import _ from "lodash";
-import { ArtistType } from "../../commands/game_options/artisttype";
 import { getMatchingGroupNames } from "../../helpers/game_utils";
-import {
-    FOREIGN_LANGUAGE_TAGS,
-    LanguageType,
-} from "../../commands/game_options/language";
-import { ShuffleType } from "../../commands/game_options/shuffle";
+import { FOREIGN_LANGUAGE_TAGS } from "../../commands/game_options/language";
 import QueriedSong from "../../interfaces/queried_song";
+import { SubunitsPreference } from "../../enums/option_types/subunit_preference";
+import { OstPreference } from "../../enums/option_types/ost_preference";
+import { ReleaseType } from "../../enums/option_types/release_type";
+import { Gender } from "../../enums/option_types/gender";
+import { ArtistType } from "../../enums/option_types/artist_type";
+import { ShuffleType } from "../../enums/option_types/shuffle_type";
+import { LanguageType } from "../../enums/option_types/language_type";
 
 async function getMockGuildPreference(): Promise<GuildPreference> {
     const guildPreference = new GuildPreference("test");
